broke out channel_controller logic from needs_team
