+++ components/needs_team/needs_team.jsx
@@ -4,7 +4,7 @@
 import $ from 'jquery';
 import PropTypes from 'prop-types';
 import React from 'react';
-import {Route, Switch, Redirect} from 'react-router-dom';
+import {Route, Switch} from 'react-router-dom';
 import iNoBounce from 'inobounce';
 
 import {getMyTeamUnreads} from 'mattermost-redux/actions/teams';
@@ -17,44 +17,18 @@ import TeamStore from 'stores/team_store.jsx';
 import UserStore from 'stores/user_store.jsx';
 import BrowserStore from 'stores/browser_store';
 import * as GlobalActions from 'actions/global_actions.jsx';
-import Pluggable from 'plugins/pluggable';
 import Constants from 'utils/constants.jsx';
 import * as UserAgent from 'utils/user_agent.jsx';
 import * as Utils from 'utils/utils.jsx';
 import {loadProfilesForSidebar} from 'actions/user_actions.jsx';
 import {checkIfMFARequired} from 'utils/route';
-import AnnouncementBar from 'components/announcement_bar';
-import DeletePostModal from 'components/delete_post_modal.jsx';
-import EditPostModal from 'components/edit_post_modal';
-import GetPostLinkModal from 'components/get_post_link_modal';
-import GetTeamInviteLinkModal from 'components/get_team_invite_link_modal';
-import GetPublicLinkModal from 'components/get_public_link_modal';
-import InviteMemberModal from 'components/invite_member_modal.jsx';
-import LeaveTeamModal from 'components/leave_team_modal.jsx';
-import LeavePrivateChannelModal from 'components/modals/leave_private_channel_modal.jsx';
-import Navbar from 'components/navbar';
-import RemovedFromChannelModal from 'components/removed_from_channel_modal.jsx';
-import ResetStatusModal from 'components/reset_status_modal';
-import ShortcutsModal from 'components/shortcuts_modal.jsx';
-import SidebarRight from 'components/sidebar_right';
-import SidebarRightMenu from 'components/sidebar_right_menu';
-import TeamSettingsModal from 'components/team_settings_modal.jsx';
-import ImportThemeModal from 'components/user_settings/import_theme_modal.jsx';
-import UserSettingsModal from 'components/user_settings/user_settings_modal.jsx';
-import WebrtcNotification from 'components/webrtc/components/webrtc_notification.jsx';
-import WebrtcSidebar from 'components/webrtc/components/webrtc_sidebar.jsx';
-import ModalController from 'components/modal_controller';
-import TeamSidebar from 'components/team_sidebar';
-import Sidebar from 'components/sidebar';
-import ChannelView from 'components/channel_view';
-import PermalinkView from 'components/permalink_view';
-import MessageIdentifierRouter from 'components/message_indentifier_router';
 import {makeAsyncComponent} from 'components/async_load';
 import loadBackstageController from 'bundle-loader?lazy!components/backstage/backstage_controller';
 
 const BackstageController = makeAsyncComponent(loadBackstageController);
 
 import store from 'stores/redux_store.jsx';
+import ChannelController from '../channel_layout/channel_controller';
 
 const dispatch = store.dispatch;
 const getState = store.getState;
@@ -214,24 +188,10 @@ export default class NeedsTeam extends React.Component {
         }
     }
 
-    toLastChannel = () => {
-        let channelName = Constants.DEFAULT_CHANNEL;
-        const team = TeamStore.getByName(this.props.match.params.team);
-        const channelId = BrowserStore.getGlobalItem(team.id);
-        const channel = ChannelStore.getChannelById(channelId);
-
-        if (channel) {
-            channelName = channel.name;
-        }
-        return `${this.props.match.url}/channels/${channelName}`;
-    }
-
     render() {
         if (this.state.team === null || this.state.finishedFetchingChannels === false) {
             return <div/>;
         }
-
-        const teamType = this.state.team ? this.state.team.type : '';
 
         return (
             <Switch>
@@ -245,62 +205,11 @@ export default class NeedsTeam extends React.Component {
                 />
                 <Route
                     render={() => (
-                        <div className='channel-view'>
-                            <AnnouncementBar/>
-                            <WebrtcNotification/>
-                            <div className='container-fluid'>
-                                <SidebarRight/>
-                                <SidebarRightMenu teamType={teamType}/>
-                                <WebrtcSidebar/>
-                                <Route component={TeamSidebar}/>
-                                <Route component={Sidebar}/>
-                                <div
-                                    id='inner-wrap-webrtc'
-                                    key='inner-wrap'
-                                    className='inner-wrap channel__wrap'
-                                >
-                                    <div className='row header'>
-                                        <div id='navbar'>
-                                            <Navbar/>
-                                        </div>
-                                    </div>
-                                    <div className='row main'>
-                                        <Switch>
-                                            <Route
-                                                path={`${this.props.match.url}/channels/:channel`}
-                                                component={ChannelView}
-                                            />
-                                            <Route
-                                                path={`${this.props.match.url}/pl/:postid`}
-                                                component={PermalinkView}
-                                            />
-                                            <Route
-                                                path={'/:team/messages/:identifier'}
-                                                component={MessageIdentifierRouter}
-                                            />
-                                            <Redirect to={this.toLastChannel()}/>
-                                        </Switch>
-                                    </div>
-                                </div>
-
-                                <Pluggable pluggableName='Root'/>
-                                <UserSettingsModal/>
-                                <GetPostLinkModal/>
-                                <GetPublicLinkModal/>
-                                <GetTeamInviteLinkModal/>
-                                <InviteMemberModal/>
-                                <LeaveTeamModal/>
-                                <ImportThemeModal/>
-                                <TeamSettingsModal/>
-                                <EditPostModal/>
-                                <DeletePostModal/>
-                                <RemovedFromChannelModal/>
-                                <ResetStatusModal/>
-                                <LeavePrivateChannelModal/>
-                                <ShortcutsModal isMac={Utils.isMac()}/>
-                                <ModalController/>
-                            </div>
-                        </div>
+                        <ChannelController
+                            teamId={this.props.match.params.team}
+                            teamType={this.state.team ? this.state.team.type : ''}
+                            matchUrl={this.props.match.url}
+                        />
                     )}
                 />
             </Switch>
